Break out goalie comparison component
