+++ components/Game/index.ts
@@ -1,10 +1,12 @@
 import DivisionStandings from './Preview/DivisionStandings';
+import GoalieComparison from './Preview/GoalieComparison';
 import SkaterComparison from './Preview/SkaterComparison';
 import TeamsBlock from './Preview/TeamsBlock';
 import TeamStats from './Preview/TeamStats';
 
 export {
   DivisionStandings,
+  GoalieComparison,
   SkaterComparison,
   TeamsBlock,
   TeamStats
